Refactoring for proper IDE imports detection
